fix(refactor to package): Refactored the code to be inside a packaged called l2cs that is then imported to support demo, train, and test scripts.
